with middleware

+++ routes/web.php
@@ -31,16 +31,14 @@ Route::group(['prefix' => 'back', 'middleware' => 'auth'], function () {
     Route::get('/category/edit', 'Admin\categoryController@edit');
 
     // Permission
+    Route::get('/permission', ['uses' => 'Admin\PermissionController@index', 'as' => 'permission-list', 'middleware' => 'permission:Permission List|All']);
+    Route::get('/permission/create', ['uses' => 'Admin\PermissionController@create', 'as' => 'permission-create', 'middleware' => 'permission:Permission List|All']);
+    Route::post('/permission/store', ['uses' => 'Admin\PermissionController@store', 'as' => 'permission-store', 'middleware' => 'permission:Permission List|All']);
+    Route::get('/permission/edit/{id}',  ['uses' => 'Admin\PermissionController@edit', 'as' => 'permission-edit', 'middleware' => 'permission:Permission List|All']);
+    Route::put('/permission/edit/{id}', ['uses' => 'Admin\PermissionController@update', 'as' => 'permission-update', 'middleware' => 'permission:Permission List|All']);
+    Route::delete('/permission/delete/{id}', ['uses' => 'Admin\PermissionController@destroy', 'as' => 'permission-delete', 'middleware' => 'permission:Permission List|All']);
 
-    Route::get('/permission', 'Admin\PermissionController@index');
-    Route::get('/permission/create', 'Admin\PermissionController@create');
-    Route::post('/permission/store', 'Admin\PermissionController@store');
-    Route::get('/permission/edit/{id}', ['uses' => 'Admin\PermissionController@edit', 'as' => 'permission-edit']);
-    Route::put('/permission/edit/{id}', ['uses' => 'Admin\PermissionController@update', 'as' => 'permission-update']);
-    Route::delete('/permission/delete/{id}', ['uses' => 'Admin\PermissionController@destroy', 'as' => 'permission-deelte']);
-
-    // Role
-
+    // roles
     Route::get('/roles', 'Admin\RoleController@index');
     Route::get('/roles/create', 'Admin\RoleController@create');
     Route::post('/roles/store', 'Admin\RoleController@store');
@@ -49,8 +47,6 @@ Route::group(['prefix' => 'back', 'middleware' => 'auth'], function () {
     Route::delete('/roles/delete/{id}', ['uses' => 'Admin\RoleController@destroy', 'as' => 'role-delete']);
 
     // author
-
-
     Route::get('/author', 'Admin\AuthorController@index');
     Route::get('/author/create', 'Admin\AuthorController@create');
     Route::post('/author/store', 'Admin\AuthorController@store');
@@ -59,35 +55,31 @@ Route::group(['prefix' => 'back', 'middleware' => 'auth'], function () {
     Route::delete('/author/delete/{id}', ['uses' => 'Admin\AuthorController@destroy', 'as' => 'author-delete']);
 
     // categroies
-
-    Route::get('/categroies', ['uses' => 'Admin\CategoryController@index', 'as' => 'category-list']);
-    Route::get('/category/create', ['uses' => 'Admin\CategoryController@create', 'as' => 'category-create']);
-    Route::post('/category/store', ['uses' => 'Admin\CategoryController@store', 'as' => 'category-store']);
-    Route::put('/category/status/{id}', ['uses' => 'Admin\CategoryController@status', 'as' => 'category-status']);
-    Route::get('/category/edit/{id}', ['uses' => 'Admin\CategoryController@edit', 'as' => 'category-edit',]);
-    Route::put('/category/update/{id}', ['uses' => 'Admin\CategoryController@update', 'as' => 'category-update']);
-    Route::delete('/category/delete/{id}', ['uses' => 'Admin\CategoryController@destroy', 'as' => 'category-delete']);
+    Route::get('/categroies', ['uses' => 'Admin\CategoryController@index', 'as' => 'category-list', 'middleware' => 'permission:Category List|All']);
+    Route::get('/category/create', ['uses' => 'Admin\CategoryController@create', 'as' => 'category-create', 'middleware' => 'permission:Category Create|All']);
+    Route::post('/category/store', ['uses' => 'Admin\CategoryController@store', 'as' => 'category-store', 'middleware' => 'permission:Category Store|All']);
+    Route::put('/category/status/{id}', ['uses' => 'Admin\CategoryController@status', 'as' => 'category-status', 'middleware' => 'permission:Category Store|All']);
+    Route::get('/category/edit/{id}', ['uses' => 'Admin\CategoryController@edit', 'as' => 'category-edit', 'middleware' => 'permission:Category edit|All']);
+    Route::put('/category/update/{id}', ['uses' => 'Admin\CategoryController@update', 'as' => 'category-update', 'middleware' => 'permission:Category update|All']);
+    Route::delete('/category/delete/{id}', ['uses' => 'Admin\CategoryController@destroy', 'as' => 'category-delete', 'middleware' => 'permission:Category delete|All']);
 
     // posts
-
-    Route::get('/posts', ['uses' => 'Admin\PostController@index', 'as' => 'post-list']);
-    Route::get('/posts/create', ['uses' => 'Admin\PostController@create', 'as' => 'post-create']);
-    Route::post('/post/store', ['uses' => 'Admin\PostController@store', 'as' => 'post-store']);
-    Route::put('/post/status/{id}', ['uses' => 'Admin\PostController@status', 'as' => 'post-status',]);
-    Route::put('/post/hot/news/{id}', ['uses' => 'Admin\PostController@hot_news', 'as' => 'hot_news-status',]);
-    Route::get('/post/edit/{id}', ['uses' => 'Admin\PostController@edit', 'as' => 'post-edit',]);
-    Route::put('/post/update/{id}', ['uses' => 'Admin\PostController@update', 'as' => 'post-update',]);
-    Route::delete('/post/delete/{id}', ['uses' => 'Admin\PostController@destroy', 'as' => 'post-delete',]);
+    Route::get('/posts', ['uses' => 'Admin\PostController@index', 'as' => 'post-list', 'middleware' => 'permission:Post List|All']);
+    Route::get('/posts/create', ['uses' => 'Admin\PostController@create', 'as' => 'post-create', 'middleware' => 'permission:Post List|All']);
+    Route::post('/post/store', ['uses' => 'Admin\PostController@store', 'as' => 'post-store', 'middleware' => 'permission:Post List|All']);
+    Route::put('/post/status/{id}', ['uses' => 'Admin\PostController@status', 'as' => 'post-status', 'middleware' => 'permission:Post List|All']);
+    Route::put('/post/hot/news/{id}', ['uses' => 'Admin\PostController@hot_news', 'as' => 'hot_news-status', 'middleware' => 'permission:Post List|All']);
+    Route::get('/post/edit/{id}', ['uses' => 'Admin\PostController@edit', 'as' => 'post-edit', 'middleware' => 'permission:Post List|All']);
+    Route::put('/post/update/{id}', ['uses' => 'Admin\PostController@update', 'as' => 'post-update', 'middleware' => 'permission:Post update|All']);
+    Route::delete('/post/delete/{id}', ['uses' => 'Admin\PostController@destroy', 'as' => 'post-delete', 'middleware' => 'permission:Post delete|All']);
 
     // comment
-
-
-    Route::get('/comment/{id}', ['uses' => 'Admin\CommentController@index', 'as' => 'comment-list',]);
-    Route::get('/comment/reply/{id}', ['uses' => 'Admin\CommentController@reply', 'as' => 'comment-view']);
-    Route::post('/comment/reply', ['uses' => 'Admin\CommentController@store', 'as' => 'comment-reply']);
-    Route::put('/comment/status/{id}', ['uses' => 'Admin\CommentController@status', 'as' => 'comment-status']);
-    Route::get('/settings', ['uses' => 'Admin\SettingController@index', 'as' => 'setting']);
-    Route::put('/settings/update', ['uses' => 'Admin\SettingController@update', 'as' => 'setting-update']);
+    Route::get('/comment/{id}', ['uses' => 'Admin\CommentController@index', 'as' => 'comment-list', 'middleware' => 'permission:Post List|All']);
+    Route::get('/comment/reply/{id}', ['uses' => 'Admin\CommentController@reply', 'as' => 'comment-view', 'middleware' => 'permission:Post List|All']);
+    Route::post('/comment/reply', ['uses' => 'Admin\CommentController@store', 'as' => 'comment-reply', 'middleware' => 'permission:Post List|All']);
+    Route::put('/comment/status/{id}', ['uses' => 'Admin\CommentController@status', 'as' => 'comment-status', 'middleware' => 'permission:Post List|All']);
+    Route::get('/settings', ['uses' => 'Admin\SettingController@index', 'as' => 'setting', 'middleware' => 'permission:Post List|All']);
+    Route::put('/settings/update', ['uses' => 'Admin\SettingController@update', 'as' => 'setting-update', 'middleware' => 'permission:Post List|All']);
 });
